+++ src/main/typescript/components/form/DepositForm.tsx
@@ -16,16 +16,16 @@
 import * as React from "react"
 import { Component } from "react"
 import Card from "./FoldableCard"
-import DepositLicenseForm from "./DepositLicenseForm"
-import PrivacySensitiveDataForm from "./PrivacySensitiveDataForm"
-import MessageForDataManagerForm from "./MessageForDataManagerForm"
-import TemporalAndSpatialCoverageForm from "./TemporalAndSpatialCoverageForm"
-import LanguageAndLiteratureSpecificMetadataForm from "./LanguageAndLiteratureSpecificMetadataForm"
-import ArchaeologySpecificMetadataForm from "./ArchaeologySpecificMetadataForm"
-import UploadTypeForm from "./UploadTypeForm"
-import LicenseAndAccessForm from "./LicenseAndAccessForm"
-import BasicInformationForm from "./BasicInformationForm"
-import DataForm from "./DataForm"
+import DepositLicenseForm from "./parts/DepositLicenseForm"
+import PrivacySensitiveDataForm from "./parts/PrivacySensitiveDataForm"
+import MessageForDataManagerForm from "./parts/MessageForDataManagerForm"
+import TemporalAndSpatialCoverageForm from "./parts/TemporalAndSpatialCoverageForm"
+import LanguageAndLiteratureSpecificMetadataForm from "./parts/LanguageAndLiteratureSpecificMetadataForm"
+import ArchaeologySpecificMetadataForm from "./parts/ArchaeologySpecificMetadataForm"
+import UploadTypeForm from "./parts/UploadTypeForm"
+import LicenseAndAccessForm from "./parts/LicenseAndAccessForm"
+import BasicInformationForm from "./parts/BasicInformationForm"
+import DataForm from "./parts/DataForm"
 import "../../../resources/css/depositForm"
 
 interface DepositFormProps {
