+++ src/main/typescript/components/form/parts/LicenseAndAccessForm.tsx
