+++ src/main/typescript/components/form/parts/TemporalAndSpatialCoverageForm.tsx
